+++ packages/react-ui/Button/Button.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { expect, test } from 'vitest'
-import { render, screen } from '@nezhos/testing/react';
+import { render, screen } from '@hugios/testing/react';
 
 import { Button } from './Button'