+++ packages/testing/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @nezhos/testing
+# @hugios/testing
 
 ## 1.1.0